TP/requirements.txt
psycopg2
python-dotenv
fire
pandas